requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
iniconfig==2.0.0
packaging==24.1
pip==23.3.2
pluggy==1.5.0
pytest~=8.2.2
setuptools==70.0.0
wheel==0.43.0
selenium~=4.21.0
Faker~=25.8.0
phone_gen~=2.4.8